// frontend/dashboard/proto/state/v1/stop_pb.d.ts
export declare enum StopStateEnum {
  /**
   * @generated from enum value: STOP_STATE_UNKNOWN = 0;
   */
  STOP_STATE_UNKNOWN = 0,

  /**
   * バーが下がってる状態
   *
   * @generated from enum value: STOP_STATE_GO = 1;
   */
  STOP_STATE_GO = 1,

  /**
   * バーが上がってる状態
   *
   * @generated from enum value: STOP_STATE_STOP = 2;
   */
  STOP_STATE_STOP = 2,
}