// frontend/dashboard/proto/state/v1/point_pb.d.ts
export declare enum PointStateEnum {
  /**
   * @generated from enum value: POINT_STATE_UNKNOWN = 0;
   */
  POINT_STATE_UNKNOWN = 0,

  /**
   * ポイントがまっすぐな状態
   *
   * @generated from enum value: POINT_STATE_NORMAL = 1;
   */
  POINT_STATE_NORMAL = 1,

  /**
   * ポイントが移動している状態
   *
   * @generated from enum value: POINT_STATE_REVERSE = 2;
   */
  POINT_STATE_REVERSE = 2,
}